Arrumamos, Produtos

+++ app/models/banco/produto.py
@@ -9,7 +9,7 @@ class Produto(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     nome = db.Column(db.String(100), nullable=False)
     descricao = db.Column(db.String(200), nullable=False)
-    preco = db.Column(db.Integer, nullable=False)
+    preco = db.Column(db.Float, nullable=False)
     quantidade = db.Column(db.Integer, nullable=False)
     arquivo = db.Column(db.String(400), nullable=False)
     vendas = db.relationship('Venda', secondary=venda_produto, backref='produto')
